+++ desafio-pix-payment.app/src/main/java/com/desafio/repository/ProofOfPaymentRepository.java
@@ -3,7 +3,7 @@ package com.desafio.repository;
 import com.desafio.model.entity.ProofOfPaymentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProofOfPaymentRepository extends JpaRepository<ProofOfPaymentEntity, Integer> {
+public interface ProofOfPaymentRepository extends JpaRepository<ProofOfPaymentEntity, Long> {
 
 
 }
